requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Diffusers with ControlNet support
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# ControlNet specific
controlnet-aux>=0.0.7  # For preprocessing (depth, canny, etc.)

# XFormers for faster inference (optional but recommended for GPU)
xformers>=0.0.22; platform_system != "Darwin"  # Skip on macOS

# Configuration
pyyaml>=6.0.0

# Model management
huggingface-hub>=0.19.0
safetensors>=0.4.0

# Optional: Better image quality
invisible-watermark>=0.2.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0